+++ examples/equal_inlet_temperature.py
@@ -9,7 +9,6 @@
 """
 import matplotlib.pyplot as plt
 import numpy as np
-from matplotlib.ticker import AutoMinorLocator
 
 import pygfunction as gt