+++ pygfunction/__init__.py
@@ -1,9 +1,11 @@
-from . import boreholes
-from . import borefield
-from . import gfunction
-from . import heat_transfer
-from . import load_aggregation
-from . import media
-from . import networks
-from . import pipes
-from . import utilities
+from . import (
+    borefield,
+    boreholes,
+    gfunction,
+    heat_transfer,
+    load_aggregation,
+    media,
+    networks,
+    pipes,
+    utilities,
+)